# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AcceptCommitmentParams": {
        "additionalProperties": false,
        "properties": {
          "accepted_terms_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Accepted Terms Hash",
            "type": "string"
          },
          "party_assignment_id": {
            "title": "Party Assignment Id",
            "type": "string"
          },
          "understanding": {
            "minLength": 10,
            "title": "Understanding",
            "type": "string"
          }
        },
        "required": [
          "party_assignment_id",
          "understanding",
          "accepted_terms_hash"
        ],
        "title": "AcceptCommitmentParams",
        "type": "object"
      },
      "ActionRequest": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "enum": [
              "create_object",
              "propose_revision",
              "accept_commitment",
              "activate_commitment",
              "confirm_adjustment",
              "confirm_closure",
              "route_feedback",
              "accept_feedback",
              "investigate_feedback",
              "confirm_decision",
              "confirm_outcome",
              "record_acceptance",
              "request_feedback_acceptance",
              "reopen_feedback",
              "revoke_assignment"
            ],
            "title": "Action Type",
            "type": "string"
          },
          "expected_versions": {
            "items": {
              "$ref": "#/components/schemas/ExpectedVersion"
            },
            "title": "Expected Versions",
            "type": "array"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateObjectParams"
              },
              {
                "$ref": "#/components/schemas/ProposeRevisionParams"
              },
              {
                "$ref": "#/components/schemas/AcceptCommitmentParams"
              },
              {
                "$ref": "#/components/schemas/ActivateCommitmentParams"
              },
              {
                "$ref": "#/components/schemas/ConfirmAdjustmentParams"
              },
              {
                "$ref": "#/components/schemas/ConfirmClosureParams"
              },
              {
                "$ref": "#/components/schemas/RouteFeedbackParams"
              },
              {
                "$ref": "#/components/schemas/RecordAcceptanceParams"
              },
              {
                "$ref": "#/components/schemas/RequestFeedbackAcceptanceParams"
              },
              {
                "$ref": "#/components/schemas/RevokeAssignmentParams"
              },
              {
                "$ref": "#/components/schemas/EmptyParams"
              }
            ],
            "title": "Params"
          },
          "reason": {
            "minLength": 5,
            "title": "Reason",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionTarget"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action_type",
          "target",
          "expected_versions",
          "idempotency_key",
          "reason",
          "params"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
...
      "ActivateCommitmentParams": {
        "additionalProperties": false,
        "properties": {
          "activation_policy_revision_id": {
            "title": "Activation Policy Revision Id",
            "type": "string"
          },
          "handshake_record_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 2,
            "title": "Handshake Record Ids",
            "type": "array"
          }
        },
        "required": [
          "handshake_record_ids",
          "activation_policy_revision_id"
        ],
        "title": "ActivateCommitmentParams",
        "type": "object"
      },
...
      "ConfirmAdjustmentParams": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/RevisionChange"
            },
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          },
          "decision_revision_id": {
            "title": "Decision Revision Id",
            "type": "string"
          },
          "feedback_revision_id": {
            "title": "Feedback Revision Id",
            "type": "string"
          }
        },
        "required": [
          "decision_revision_id",
          "changes",
          "feedback_revision_id"
        ],
        "title": "ConfirmAdjustmentParams",
        "type": "object"
      },
      "ConfirmClosureParams": {
        "additionalProperties": false,
        "properties": {
          "acceptance_record_id": {
            "title": "Acceptance Record Id",
            "type": "string"
          },
          "closure_evidence_revision_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Closure Evidence Revision Ids",
            "type": "array"
          },
          "closure_note": {
            "minLength": 10,
            "title": "Closure Note",
            "type": "string"
          },
          "disposition": {
            "enum": [
              "resolved",
              "dismissed",
              "no_change"
            ],
            "title": "Disposition",
            "type": "string"
          },
          "resolution_decision_revision_id": {
            "title": "Resolution Decision Revision Id",
            "type": "string"
          }
        },
        "required": [
          "acceptance_record_id",
          "resolution_decision_revision_id",
          "closure_evidence_revision_ids",
          "disposition",
          "closure_note"
        ],
        "title": "ConfirmClosureParams",
        "type": "object"
      },
      "ContextRequest": {
        "additionalProperties": false,
        "properties": {
          "known_at": {
            "format": "date-time",
            "title": "Known At",
            "type": "string"
          },
          "object_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Object Ids",
            "type": "array"
          },
          "valid_at": {
            "format": "date-time",
            "title": "Valid At",
            "type": "string"
          }
        },
        "required": [
          "object_ids",
          "valid_at",
          "known_at"
        ],
        "title": "ContextRequest",
        "type": "object"
      },
      "CreateObjectParams": {
        "additionalProperties": false,
        "properties": {
          "domain_id": {
            "title": "Domain Id",
            "type": "string"
          },
          "object_type": {
            "enum": [
              "CompanyOutcome",
              "BusinessCommitment",
              "ExecutionCommitment",
              "FeedbackThread",
              "ManagementAdjustment",
              "Decision",
              "MetricObservation"
            ],
            "title": "Object Type",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommitmentPayload"
              },
              {
                "$ref": "#/components/schemas/FeedbackPayload"
              },
              {
                "$ref": "#/components/schemas/DecisionPayload"
              },
              {
                "$ref": "#/components/schemas/AdjustmentPayload"
              },
              {
                "$ref": "#/components/schemas/ObservationPayload"
              },
              {
                "$ref": "#/components/schemas/OutcomePayload"
              }
            ],
            "title": "Payload"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          }
        },
        "required": [
          "object_type",
          "domain_id",
          "payload"
        ],
        "title": "CreateObjectParams",
        "type": "object"
      },
...
      "EmptyParams": {
        "additionalProperties": false,
        "properties": {},
        "title": "EmptyParams",
        "type": "object"
      },
      "EvidenceRequest": {
        "additionalProperties": false,
        "properties": {
          "content_base64": {
            "maxLength": 2796204,
            "minLength": 1,
            "title": "Content Base64",
            "type": "string"
          },
          "domain_id": {
            "format": "uuid",
            "title": "Domain Id",
            "type": "string"
          },
          "media_type": {
            "default": "application/octet-stream",
            "maxLength": 120,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9!#$&^_.+-]+/[A-Za-z0-9!#$&^_.+-]+$",
            "title": "Media Type",
            "type": "string"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "domain_id",
          "title",
          "content_base64"
        ],
        "title": "EvidenceRequest",
        "type": "object"
      },
      "ExpectedVersion": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          }
        },
        "required": [
          "object_id",
          "expected_version"
        ],
        "title": "ExpectedVersion",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProposeRevisionParams": {
        "additionalProperties": false,
        "properties": {
          "bundle_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Id"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommitmentPayload"
              },
              {
                "$ref": "#/components/schemas/FeedbackPayload"
              },
              {
                "$ref": "#/components/schemas/DecisionPayload"
              },
              {
                "$ref": "#/components/schemas/AdjustmentPayload"
              },
              {
                "$ref": "#/components/schemas/ObservationPayload"
              },
              {
                "$ref": "#/components/schemas/OutcomePayload"
              }
            ],
            "title": "Payload"
          }
        },
        "required": [
          "payload"
        ],
        "title": "ProposeRevisionParams",
        "type": "object"
      },
...
      "RequestFeedbackAcceptanceParams": {
        "additionalProperties": false,
        "properties": {
          "decision_revision_id": {
            "title": "Decision Revision Id",
            "type": "string"
          }
        },
        "required": [
          "decision_revision_id"
        ],
        "title": "RequestFeedbackAcceptanceParams",
        "type": "object"
      },
...
      "RouteFeedbackParams": {
        "additionalProperties": false,
        "properties": {
          "responsible_assignment_id": {
            "title": "Responsible Assignment Id",
            "type": "string"
          }
        },
        "required": [
          "responsible_assignment_id"
        ],
        "title": "RouteFeedbackParams",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Standalone governed Memory Service with a native API and a temporary Clark GraphKnowledge compatibility fa\u00e7ade.",
    "title": "TKOS Memory Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/action-receipts/{receipt_id}": {
      "get": {
        "operationId": "receipt_get_v1_action_receipts__receipt_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "receipt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Receipt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipt Get",
        "tags": [
          "governed-runtime"
        ]
      }
    },
    "/v1/actions": {
      "post": {
        "operationId": "action_v1_actions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action",
        "tags": [
          "governed-runtime"
        ]
      }
    },
    "/v1/actions/prepare": {
      "post": {
        "operationId": "prepare_v1_actions_prepare_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare",
        "tags": [
          "governed-runtime"
        ]
      }
    },
    "/v1/context-packs": {
      "post": {
        "operationId": "context_create_v1_context_packs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Context Create",
        "tags": [
          "governed-runtime"
        ]
      }
    },
    "/v1/context-packs/{snapshot_id}": {
      "get": {
        "operationId": "snapshot_get_v1_context_packs__snapshot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Snapshot Get",
        "tags": [
          "governed-runtime"
        ]
      }
    },
    "/v1/evidence-assets": {
      "post": {
        "operationId": "evidence_create_v1_evidence_assets_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence Create",
        "tags": [
          "governed-runtime"
        ]
      }
    },
    "/v1/evidence-assets/{object_id}/revisions/{revision_id}": {
      "get": {
        "operationId": "evidence_download_v1_evidence_assets__object_id__revisions__revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence Download",
        "tags": [
          "governed-runtime"
        ]
      }
    },
...
    "/v1/objects/{object_id}": {
      "get": {
        "operationId": "object_get_v1_objects__object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Object Get",
        "tags": [
          "governed-runtime"
        ]
      }
    },
    "/v1/objects/{object_id}/revisions/{revision_id}": {
      "get": {
        "operationId": "revision_get_v1_objects__object_id__revisions__revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revision Get",
        "tags": [
          "governed-runtime"
        ]
      }
    }
  }
}